add tests around fileset aggregation functions

+++ hs_core/views/utils.py
@@ -1034,15 +1034,15 @@ def unzip_file(user, res_id, zip_with_rel_path, bool_remove_original,
             unzip_path = istorage.unzip(zip_with_full_path)
             link_irods_folder_to_django(resource, istorage, unzip_path, auto_aggregate)
 
-            if bool_remove_original:
-                zip_with_rel_path = zip_with_rel_path.split("contents/", 1)[1]
-                delete_resource_file(res_id, zip_with_rel_path, user)
-
     except Exception:
         logger.exception("failed to unzip")
         if unzip_path and istorage.exists(unzip_path):
             istorage.delete(unzip_path)
         raise
+
+    if bool_remove_original and not ingest_metadata:  # ingest_metadata deletes the zip by default
+        zip_with_rel_path = zip_with_rel_path.split("contents/", 1)[1]
+        delete_resource_file(res_id, zip_with_rel_path, user)
 
     # TODO: should check can_be_public_or_discoverable here
     resource.refresh_from_db()
